Move update logic to edit spending cap modal
